Full 2

+++ DevideManagement/FormDevice.cs
@@ -57,45 +57,63 @@ namespace DeviceManagement
 
         private void LoadTable()
         {
-            List<DeviceModel> data = DevicePresenter.GetCurrentPage();
-            if (data.Count == 0) return;
-            List<dynamic> devices = new List<dynamic>();
-            foreach(var item in data)
+            try
             {
-                devices.Add(new
+                // Load device list
+                List<DeviceModel> data = DevicePresenter.GetCurrentPage();
+                btnPrePage.Enabled = DevicePresenter.HasPreviousPage();
+                btnNextPage.Enabled = DevicePresenter.HasNextPage();
+                if (data.Count == 0) return;
+                List<dynamic> devices = new List<dynamic>();
+                foreach (var item in data)
                 {
-                    Id = item.Id,
-                    Name = item.Name,
-                    CategoryName = item.CategoryName,
-                    ImageView = item.ImageView,
-                    Description = item.Description,
-                    BoughtDate = item.BoughtDate,
-                    WarrantyDate = item.WarrantyDate,
-                    StatusName = item.StatusName
-                });
+                    devices.Add(new
+                    {
+                        Id = item.Id,
+                        Name = item.Name,
+                        CategoryName = item.CategoryName,
+                        ImageView = item.ImageView,
+                        Description = item.Description,
+                        BoughtDate = item.BoughtDate,
+                        WarrantyDate = item.WarrantyDate,
+                        StatusName = item.StatusName
+                    });
+                }
+                dgvDevice.DataSource = devices;
+                // Hide some column
+                for (int i = 2; i <= 7; i++) dgvDevice.Columns[i].Visible = false;
+
+                // Data bindings
+                txtDeviceId.DataBindings.Clear();
+                txtDeviceName.DataBindings.Clear();
+                cbCategory.DataBindings.Clear();
+                pbImage.DataBindings.Clear();
+                txtDeviceDescription.DataBindings.Clear();
+                lbBoughtDate.DataBindings.Clear();
+                lbWarrantyDate.DataBindings.Clear();
+                lbStatus.DataBindings.Clear();
+                txtDeviceId.DataBindings.Add("Text", devices, "Id");
+                txtDeviceName.DataBindings.Add("Text", devices, "Name");
+                cbCategory.DataBindings.Add("Text", devices, "CategoryName");
+                pbImage.DataBindings.Add(new Binding("Image", devices, "ImageView", true));
+                txtDeviceDescription.DataBindings.Add("Text", devices, "Description");
+                lbBoughtDate.DataBindings.Add("Text", devices, "BoughtDate");
+                lbWarrantyDate.DataBindings.Add("Text", devices, "WarrantyDate");
+                lbStatus.DataBindings.Add("Text", devices, "StatusName");
+                dgvDevice.ClearSelection();
+
+                // Load category list
+                cbCategory.Items.Clear();
+                List<CategoryModel> categories = DevicePresenter.GetCategories();
+                foreach (var category in categories)
+                {
+                    cbCategory.Items.Add(category.Name);
+                }
             }
-            dgvDevice.DataSource = devices;
-            // hide some column
-            for (int i = 2; i <= 7; i++) dgvDevice.Columns[i].Visible = false;
-            txtDeviceId.DataBindings.Clear();
-            txtDeviceName.DataBindings.Clear();
-            cbCategory.DataBindings.Clear();
-            pbImage.DataBindings.Clear();
-            txtDeviceDescription.DataBindings.Clear();
-            txtBoughtDate.DataBindings.Clear();
-            txtWarrantyDate.DataBindings.Clear();
-            lbStatus.DataBindings.Clear();
-            txtDeviceId.DataBindings.Add("Text", devices, "Id");
-            txtDeviceName.DataBindings.Add("Text", devices, "Name");
-            cbCategory.DataBindings.Add("Text", devices, "CategoryName");
-            pbImage.DataBindings.Add(new Binding("Image", devices, "ImageView", true));
-            txtDeviceDescription.DataBindings.Add("Text", devices, "Description");
-            txtBoughtDate.DataBindings.Add("Text", devices, "BoughtDate");
-            txtWarrantyDate.DataBindings.Add("Text", devices, "WarrantyDate");
-            lbStatus.DataBindings.Add("Text", devices, "StatusName");
-            dgvDevice.ClearSelection();
-            btnPrePage.Enabled = DevicePresenter.HasPreviousPage();
-            btnNextPage.Enabled = DevicePresenter.HasNextPage();
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message, "Load failed");
+            }
         }
 
         private void FrmDevice_Load(object sender, EventArgs e)
@@ -121,58 +139,14 @@ namespace DeviceManagement
             txtDeviceDescription.Text = string.Empty;
             txtBoughtDate.Text = string.Empty;
             txtWarrantyDate.Text = string.Empty;
-            lbStatusActive.Visible = true;
+            lbStatus.Text = "DEVICE_ACTIVE";
             pbImage.Image = null;
         }
 
         private void btnAdd_Click(object sender, EventArgs e)
         {
             if (txtDeviceId.Text.Length != 0) return;
-            if (txtDeviceName.Text.Trim().Length == 0)
-            {
-                MessageBox.Show("Device name must be not empty.", "Error");
-                return;
-            }
-            if (txtDeviceName.Text.Trim().Length > 50)
-            {
-                MessageBox.Show("Device name must be not greater than 50 chars.", "Error");
-                return;
-            }
-            if (txtDeviceDescription.Text.Trim().Length > 300)
-            {
-                MessageBox.Show("Device description must be not greater than 300 chars.", "Error");
-                return;
-            }
-            DateTime boughtDate = default;
-            try
-            {
-                if (txtBoughtDate.Text.Trim().Length > 0) 
-                { 
-                    boughtDate = DateTime.Parse(txtBoughtDate.Text.Trim());
-                }
-            } catch
-            {
-                MessageBox.Show("Bought date is wrong format.", "Error");
-                return;
-            }
-            DateTime warrantyDate = default;
-            try
-            {
-                if (txtWarrantyDate.Text.Trim().Length > 0)
-                {
-                    warrantyDate = DateTime.Parse(txtWarrantyDate.Text.Trim());
-                }
-            }
-            catch
-            {
-                MessageBox.Show("Warranty date is wrong format.", "Error");
-                return;
-            }
-            if (boughtDate != default && warrantyDate != default && DateTime.Compare(boughtDate, warrantyDate) > 0) 
-            {
-                MessageBox.Show("Warranty date must be later than bought date.", "Error");
-                return;
-            }
+            if (!isValidInput()) return;
             try
             {
                 DevicePresenter.AddDeviceModel();
@@ -180,51 +154,16 @@ namespace DeviceManagement
                 LoadTable();
                 MessageBox.Show("Add successful.", "Information");
             }
-            catch
+            catch (Exception ex)
             {
-                MessageBox.Show("Add failed", "Information");
+                MessageBox.Show(ex.Message, "Add failed");
             }
         }
 
         private void btnUpdate_Click(object sender, EventArgs e)
         {
             if (txtDeviceId.Text.Length == 0) return;
-            if (txtDeviceName.Text.Trim().Length == 0)
-            {
-                MessageBox.Show("Device name must be not empty.", "Error");
-                return;
-            }
-            DateTime boughtDate = default;
-            try
-            {
-                if (txtBoughtDate.Text.Trim().Length > 0)
-                {
-                    boughtDate = DateTime.Parse(txtBoughtDate.Text.Trim());
-                }
-            }
-            catch
-            {
-                MessageBox.Show("Bought date is wrong format.", "Error");
-                return;
-            }
-            DateTime warrantyDate = default;
-            try
-            {
-                if (txtWarrantyDate.Text.Trim().Length > 0)
-                {
-                    warrantyDate = DateTime.Parse(txtWarrantyDate.Text.Trim());
-                }
-            }
-            catch
-            {
-                MessageBox.Show("Warranty date is wrong format.", "Error");
-                return;
-            }
-            if (boughtDate != default && warrantyDate != default && DateTime.Compare(boughtDate, warrantyDate) > 0)
-            {
-                MessageBox.Show("Warranty date must be later than bought date.", "Error");
-                return;
-            }
+            if (!isValidInput()) return;
             try
             {
                 DevicePresenter.UpdateDeviceModel();
@@ -232,10 +171,11 @@ namespace DeviceManagement
                 LoadTable();
                 MessageBox.Show("Update successful.", "Information");
             }
-            catch
+            catch (Exception ex)
             {
-                MessageBox.Show("Update failed", "Information");
-            } finally 
+                MessageBox.Show(ex.Message, "Update failed");
+            }
+            finally 
             {
                 isChangeImage = false;
             }
@@ -244,15 +184,18 @@ namespace DeviceManagement
         private void btnDelete_Click(object sender, EventArgs e)
         {
             if (txtDeviceId.Text.Length == 0) return;
+            DialogResult dr = MessageBox.Show("Are you sure to delete this device?",
+                "Delete Device", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Information);
+            if (dr != DialogResult.Yes) return;
             try
             {
                 DevicePresenter.DeleteDeviceModel();
                 LoadTable();
                 MessageBox.Show("Delete successful.", "Information");
             }
-            catch
+            catch (Exception ex)
             {
-                MessageBox.Show("Delete failed", "Information");
+                MessageBox.Show(ex.Message, "Delete failed");
             }
         }
 
@@ -283,12 +226,6 @@ namespace DeviceManagement
 
         private void dgvDevice_SelectionChanged(object sender, EventArgs e)
         {
-            cbCategory.Items.Clear();
-            List<CategoryModel> categories = DevicePresenter.GetCategories();
-            foreach (var category in categories)
-            {
-                cbCategory.Items.Add(category.Name);
-            }
             isChangeImage = false;
         }
 
@@ -302,9 +239,9 @@ namespace DeviceManagement
                 LoadTable();
                 MessageBox.Show("Remove successful.", "Information");
             }
-            catch
+            catch (Exception ex)
             {
-                MessageBox.Show("Remove failed", "Information");
+                MessageBox.Show(ex.Message, "Remove failed");
             }
         }
 
@@ -314,14 +251,9 @@ namespace DeviceManagement
             lbStatusInactive.Visible = !lbStatusActive.Visible;
         }
 
-        private void txtBoughtDate_Enter(object sender, EventArgs e)
-        {
-            txtBoughtDate.DataBindings.Clear();
-            txtWarrantyDate.DataBindings.Clear();
-        }
-
         private void btnHistory_Click(object sender, EventArgs e)
         {
+            if (dgvDevice.SelectedRows.Count == 0) return;
             new FormRequest(Id, UserModel, ref lbStatus).ShowDialog();
             LoadTable();
         }
@@ -337,6 +269,82 @@ namespace DeviceManagement
             {
                 LoadTable();
             }
+        }
+
+        private void lbBoughtDate_TextChanged(object sender, EventArgs e)
+        {
+            string date = lbBoughtDate.Text.Substring(0, lbBoughtDate.Text.IndexOf(" "));
+            txtBoughtDate.Text = date;
+        }
+
+        private void lbWarrantyDate_TextChanged(object sender, EventArgs e)
+        {
+            string date = lbWarrantyDate.Text.Substring(0, lbWarrantyDate.Text.IndexOf(" "));
+            txtWarrantyDate.Text = date;
+        }
+
+        private bool isValidInput()
+        {
+            // Validate device name
+            if (string.IsNullOrWhiteSpace(txtDeviceName.Text))
+            {
+                MessageBox.Show("Device name must be not empty.", "Error");
+                return false;
+            }
+            if (txtDeviceName.Text.Trim().Length > 50)
+            {
+                MessageBox.Show("Device name must be not greater than 50 chars.", "Error");
+                return false;
+            }
+
+            // Validate device description
+            if (txtDeviceDescription.Text.Trim().Length > 300)
+            {
+                MessageBox.Show("Device description must be not greater than 300 chars.", "Error");
+                return false;
+            }
+
+            // Validate bought date
+            if (string.IsNullOrWhiteSpace(txtBoughtDate.Text))
+            {
+                MessageBox.Show("Bought date must be not empty.", "Error");
+                return false;
+            }
+            DateTime boughtDate = default;
+            try
+            {
+                boughtDate = DateTime.Parse(txtBoughtDate.Text.Trim());
+            }
+            catch
+            {
+                MessageBox.Show("Bought date is wrong format.", "Error");
+                return false;
+            }
+
+            // Validate warranty date
+            if (string.IsNullOrWhiteSpace(txtWarrantyDate.Text))
+            {
+                MessageBox.Show("Warranty date must be not empty.", "Error");
+                return false;
+            }
+            DateTime warrantyDate = default;
+            try
+            {
+                warrantyDate = DateTime.Parse(txtWarrantyDate.Text.Trim());
+            }
+            catch
+            {
+                MessageBox.Show("Warranty date is wrong format.", "Error");
+                return false;
+            }
+            if (boughtDate != default && warrantyDate != default && DateTime.Compare(boughtDate, warrantyDate) > 0)
+            {
+                MessageBox.Show("Warranty date must be later than bought date.", "Error");
+                return false;
+            }
+
+            // No error
+            return true;
         }
     }
 }
