+++ DevideManagement/FormWorker.cs
@@ -28,20 +28,29 @@ namespace DeviceManagement
 
         private void LoadTable()
         {
-            List<UserModel> list = WorkerPresenter.GetCurrentPage();
-            dgvUser.DataSource = list;
-            dgvUser.Columns[2].Visible = false;
-            dgvUser.Columns[3].Visible = false;
-            dgvUser.Columns[4].Visible = false;
-            txtUserID.DataBindings.Clear();
-            txtUsername.DataBindings.Clear();
-            txtPassword.DataBindings.Clear();
-            txtUserID.DataBindings.Add("Text", list, "id");
-            txtUsername.DataBindings.Add("Text", list, "username");
-            txtPassword.DataBindings.Add("Text", list, "password");
-            dgvUser.ClearSelection();
-            btnPrePage.Enabled = WorkerPresenter.HasPreviousPage();
-            btnNextPage.Enabled = WorkerPresenter.HasNextPage();
+            try
+            {
+                List<UserModel> list = WorkerPresenter.GetCurrentPage();
+                btnPrePage.Enabled = WorkerPresenter.HasPreviousPage();
+                btnNextPage.Enabled = WorkerPresenter.HasNextPage();
+                if (list.Count == 0) return;
+                dgvUser.DataSource = list;
+                // Hide some columns
+                dgvUser.Columns[2].Visible = false;
+                dgvUser.Columns[3].Visible = false;
+                dgvUser.Columns[4].Visible = false;
+                // Data bindings
+                txtUserID.DataBindings.Clear();
+                txtUsername.DataBindings.Clear();
+                txtPassword.DataBindings.Clear();
+                txtUserID.DataBindings.Add("Text", list, "id");
+                txtUsername.DataBindings.Add("Text", list, "username");
+                txtPassword.DataBindings.Add("Text", list, "password");
+                dgvUser.ClearSelection();
+            } catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message, "Load failed");
+            }
         }
 
         private void ckShowPassword_CheckedChanged(object sender, EventArgs e)
@@ -67,31 +76,10 @@ namespace DeviceManagement
         private void btnAdd_Click(object sender, EventArgs e)
         {
             if (txtUserID.Text.Length != 0) return;
-            string username = txtUsername.Text.Trim();
-            if (username.Length == 0)
-            {
-                MessageBox.Show("Username must be not empty.", "Error");
-                return;
-            }
-            if (username.Length > 50)
-            {
-                MessageBox.Show("Username must be not greater than 50 chars.", "Error");
-                return;
-            }
+            if (!IsValidInput()) return;
             if (WorkerPresenter.IsExistUsername())
             {
                 MessageBox.Show("This username has existed.", "Error");
-                return;
-            }
-            string password = txtPassword.Text.Trim();
-            if (password.Length == 0)
-            {
-                MessageBox.Show("Password must be not empty.", "Error");
-                return;
-            }
-            if (password.Length > 50)
-            {
-                MessageBox.Show("Password must be not greater than 50 chars.", "Error");
                 return;
             }
             try
@@ -101,45 +89,43 @@ namespace DeviceManagement
                 LoadTable();
                 MessageBox.Show("Add successful.", "Information");
             }
-            catch
+            catch (Exception ex)
             {
-                MessageBox.Show("Add failed", "Information");
+                MessageBox.Show(ex.Message, "Add failed");
             }
         }
 
         private void btnUpdate_Click(object sender, EventArgs e)
         {
             if (txtUserID.Text.Length == 0) return;
-            string password = txtPassword.Text.Trim();
-            if (password.Length == 0)
-            {
-                MessageBox.Show("Password must be not empty.", "Error");
-                return;
-            }
+            if (!IsValidInput()) return;
             try
             {
                 WorkerPresenter.UpdateUserModel();
                 LoadTable();
                 MessageBox.Show("Update successful.", "Information");
             }
-            catch
+            catch (Exception ex)
             {
-                MessageBox.Show("Update failed", "Information");
+                MessageBox.Show(ex.Message, "Update failed");
             }
         }
 
         private void btnDelete_Click(object sender, EventArgs e)
         {
             if (txtUserID.Text.Length == 0) return;
+            DialogResult dr = MessageBox.Show("Are you sure to delete this worker?",
+                "Delete Worker", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Information);
+            if (dr != DialogResult.Yes) return;
             try
             {
                 WorkerPresenter.DeleteUserModel();
                 LoadTable();
                 MessageBox.Show("Delete successful.", "Information");
             }
-            catch
+            catch (Exception ex)
             {
-                MessageBox.Show("Delete failed", "Information");
+                MessageBox.Show(ex.Message, "Delete failed");
             }
         }
 
@@ -169,14 +155,52 @@ namespace DeviceManagement
         private void btnSearch_Click(object sender, EventArgs e)
         {
             WorkerPresenter.GoToFirstPage();
-            if (WorkerPresenter.GetCurrentPage().Count == 0)
+            try
             {
-                MessageBox.Show("No workers are matched.", "Information");
+                int records = WorkerPresenter.GetCurrentPage().Count;
+                if (records == 0)
+                {
+                    MessageBox.Show("No workers are matched.", "Information");
+                }
+                else
+                {
+                    LoadTable();
+                }
             }
-            else
+            catch (Exception ex)
             {
-                LoadTable();
+                MessageBox.Show(ex.Message, "Search failed");
             }
+        }
+
+        private bool IsValidInput()
+        {
+            // Validate username
+            if (string.IsNullOrWhiteSpace(txtUsername.Text))
+            {
+                MessageBox.Show("Username must be not empty.", "Error");
+                return false;
+            }
+            if (txtUsername.Text.Length > 50)
+            {
+                MessageBox.Show("Username must be not greater than 50 chars.", "Error");
+                return false;
+            }
+
+            // Validate password
+            if (string.IsNullOrWhiteSpace(txtPassword.Text))
+            {
+                MessageBox.Show("Password must be not empty.", "Error");
+                return false;
+            }
+            if (txtPassword.Text.Length > 50)
+            {
+                MessageBox.Show("Password must be not greater than 50 chars.", "Error");
+                return false;
+            }
+
+            // No error
+            return true;
         }
     }
 }
